clean(mxfp_moe): dedup shared helpers into common; trim dead code
Consolidate helpers duplicated across gemm1/gemm2 into mxfp4_gemm_common
and drop dead code left by the layout-API port. Pure relocation/trim; no
codegen change (verified cold: variants 6/6 + a4w4/a8w4 FP4-M atomic +
FP4-L reduce pass).

- Hoist _scale_mma_atoms (+_A_ELEM) to common; gemm2 now calls
  _scale_mma_atoms("fp4") instead of its inline fp4xfp4 atom dict.
- Hoist _global_i32_buffer_view/_global_i32_buffer_tiles to common;
  remove gemm1's nested copies and gemm2's module-level copies.
- Remove gemm1's local _udiv/_umod; import from common (gemm2 already did).
- Drop gemm1 mfma_cluster's dead `init` param (del init) and the two
  init=... call args.
- Trim redundant fx.Int32(bx_i32) re-wrap; drop now-unused Float4/Float8
  imports from gemm1/gemm2.
